move client types into new module
